move running_node to its file, get rid of gigantic node.rs
